requirements.txt
Django==1.7.6
django-cors-headers==1.0.0
djangorestframework==3.1.0
